Refactor: Divide ClientsPage components in Actions, Header and Table components #6
